+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Issue [#24](https://github.com/m-marini/qucomp/issues/24): Add qu language compiler
 - Issue [#26](https://github.com/m-marini/qucomp/issues/26): Add matrix qu language
-- Issue [#28](https://github.com/m-marini/qucomp/issues/28): Add parameter list
+- Issue [#27](https://github.com/m-marini/qucomp/issues/27): Add parameter list
+- Issue [#28](https://github.com/m-marini/qucomp/issues/28): Add variables panel
 
 ## [0.2.0] 2025-05-15